priors are a kind of variable
